requirements.txt
asgiref==3.9.2
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
ConfigArgParse==1.7.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
dnspython==2.8.0
Faker==19.6.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
geographiclib==2.1
geopy==2.3.0
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.4
h11==0.16.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.40.4
locust-cloud==1.26.3
MarkupSafe==3.0.2
msgpack==1.1.1
numpy==1.26.0
packaging==25.0
pika==1.3.2
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
psutil==7.0.0
psycopg2==2.9.10
pycparser==2.23
Pygments==2.19.2
pymongo==4.15.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
pywin32==311
pyzmq==27.1.0
reportlab==4.4.3
requests==2.31.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==6.0
zope.interface==8.0
